requirements.txt
# Core dependencies
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-text-splitters==0.0.1
langgraph==0.0.10

# Vector store and embeddings
chromadb==0.4.18
rank-bm25==0.2.2

# Document processing
docling==0.1.0
python-docx==1.0.1
PyPDF2==3.0.1

# API integrations
openai==1.3.7
cohere==4.37
backoff==2.2.1

# Monitoring and metrics
prometheus-client==0.19.0
structlog==24.1.0

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Configuration and validation
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML==6.0.1

# Type checking
mypy==1.7.1
types-PyYAML==6.0.12.12
types-python-dateutil==2.8.19.14

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings==0.24.0
